various additions
- Changed module setting *singular* to false. This enables you load the
module as a new instance with `$nav =
$modules->MarkupSimpleNavigation`. This changes behavior for hooks
attached to such an instance as it will only be called for this
instance. This allows you to create multiple instances of the module to
create different navigations with separate hooks.
- Added support for `Page::listable`. So you can exclude listing of
pages in runtime by modifying this page permission. This can be done
using a system hook on `Page::listable` and set the hook event to
return `false`.
- Added new option **date_format** for `created` and `modified` native
page properties
- Added new option **code_formatting**, to enable or disable code
formatting (indentation and newlines) of output markup
- Added new option **debug** that will output rendertime and selector
infos as commented inline html